-fix cost bug (it should be a float and not an int) -add a test to prove the fix -add another test using string values instead of numbers

+++ app/views/expense_view.py
@@ -22,7 +22,7 @@ class ExpenseView:
             raise ValueError(self.EMPTY_VALUE_ERROR_MESSAGE)
         expense_cost = input("Enter how much you've spent (£): ")
         try:
-            expense_cost = int(expense_cost)
+            expense_cost = float(expense_cost)
         except ValueError:
             raise ValueError(self.WRONG_TYPE_MESSAGE)
